% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_save.R
\name{cache_save}
\alias{cache_save}
\title{Save a cache file (name generated with \code{code_name()}}
\usage{
cache_save(object, filename)
}
\arguments{
\item{object}{The object to cache (i.e. the value of the evaluating function).}

\item{filename}{The cache file name (i.e. use the one generated by cache_name, see examples).}
}
\description{
Saves the object as an .RDS file with the filename
}
\details{
See the examples for \code{\link{cache_name}()}.
}
