% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.list.R
\name{check}
\alias{check}
\title{Checking the object for appropriate form.}
\usage{
check(object)
}
\arguments{
\item{object}{The object to be checked.}
}
\value{
The tables generated.

# Check a data.list (see \code{?\link{check.data.list}})
check(Dbuilding)
}
\description{
Checking the object for appropriate form.
}
\details{
Prints on table form the result of the check.
}
