% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.list.R
\name{==.data.list}
\alias{==.data.list}
\title{Determine if two data.lists are identical}
\usage{
\method{==}{data.list}(x, y)
}
\arguments{
\item{x}{first data.list}

\item{y}{second data.list}
}
\value{
logical
}
\description{
Compare two data.lists
}
\details{
Returns TRUE if the two data.lists are fully identical, so all data, order of variables etc. must be fully identical
}
\examples{

Dbuilding == Dbuilding

D <- Dbuilding
D$Ta$k2[1] <- NA
Dbuilding == D

D <- Dbuilding
names(D)[5] <- "I"
names(D)[6] <- "Ta"
Dbuilding == D


}
