% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenlist.R
\name{flattenlist}
\alias{flattenlist}
\title{Flattens list}
\usage{
flattenlist(x)
}
\arguments{
\item{x}{List to flatten.}
}
\value{
A flatten list
}
\description{
Flattens list in a single list of data.frames
}
\details{
Flattens list. Can maybe be made better. It might end up copying data in
memory!? It might change the order of the elements.
}
