% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagdl.R
\name{lagdl}
\alias{lagdl}
\title{Lagging which returns a data.list}
\usage{
lagdl(DL, lagseq)
}
\arguments{
\item{DL}{The data.list to be lagged.}

\item{lagseq}{The integer(s) setting the lag steps.}
}
\value{
A data.list.
}
\description{
Lagging by shifting the values back or fourth always returning a data.list.
}
\details{
This function lags (shifts) the values of the vector. A data.list is always returned with each data.frame lagged with \code{lagdf}.
}
\examples{
# The values are simply shifted in each data.frame with lagdf

}
