% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\name{lapply_cbind}
\alias{lapply_cbind}
\title{Helper which does lapply and then cbind}
\usage{
lapply_cbind(X, FUN)
}
\arguments{
\item{X}{object to apply on}

\item{FUN}{function to apply}
}
\description{
Helper which does lapply and then cbind
}
