% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\name{lapply_rbind}
\alias{lapply_rbind}
\title{Helper which does lapply and then rbind}
\usage{
lapply_rbind(X, FUN)
}
\arguments{
\item{X}{object to apply on}

\item{FUN}{function to apply}
}
\description{
Helper which does lapply and then rbind
}
