% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply.R
\name{lapply_rbind_df}
\alias{lapply_rbind_df}
\title{Helper which does lapply, rbind and then as.data.frame}
\usage{
lapply_rbind_df(X, FUN)
}
\arguments{
\item{X}{object to apply on}

\item{FUN}{function to apply}
}
\description{
Helper which does lapply, rbind and then as.data.frame
}
