% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lp_vector}
\alias{lp_vector}
\title{First-order low-pass filtering}
\usage{
lp_vector(x, a1)
}
\arguments{
\item{x}{vector.}

\item{a1}{The low-pass filter coefficient.}
}
\value{
The low-pass filtered vector
}
\description{
First-order low-pass filtering of a time series vector.
}
\details{
This function applies a first order unity gain low-pass filter vector.

The \code{lp_vector_cpp} function does the same much faster.
}
