% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lp_vector_cpp}
\alias{lp_vector_cpp}
\title{Low pass filtering of a vector.}
\arguments{
\item{x}{A numeric vector}

\item{a1}{the first order low-pass filter coefficient}
}
\description{
This function returns a vector which is x through a unity gain first-order low-pass filter.
}
