% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ones.R
\name{one}
\alias{one}
\title{Create ones for model input intercept}
\usage{
one()
}
\value{
A data.frame of ones
}
\description{
Returns a data.frame of ones which can be used in forecast model inputs
}
\details{
The function returns ones which can be used to generate ones, e.g. to be used as a intercept for a model.

See vignettes 'setup-data' and 'setup-and-use-model'.
}
\examples{

# A model
model <- forecastmodel$new()
# Use the function in the input definition
model$add_inputs(mu = "one()")
# Set the forecast horizons
model$kseq <- 1:4
# During the transformation stage the one will be generated for the horizons
model$transform_data(subset(Dbuilding, 1:7))

}
