% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineforecast-package.R
\docType{package}
\name{onlineforecast-package}
\alias{onlineforecast}
\alias{onlineforecast-package}
\title{onlineforecast: Forecast Modelling for Online Applications}
\description{
A framework for fitting adaptive forecasting models. Provides a way to use forecasts as input to models, e.g. weather forecasts for energy related forecasting. The models can be fitted recursively and can easily be setup for updating parameters when new data arrives. See the included vignettes, the website \url{https://onlineforecasting.org} and the paper "onlineforecast: An R package for adaptive and recursive forecasting" \url{https://journal.r-project.org/articles/RJ-2023-031/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://onlineforecasting.org}
  \item Report bugs at \url{https://lab.compute.dtu.dk/packages/onlineforecast/-/issues}
}

}
\author{
\strong{Maintainer}: Peder Bacher \email{pbac@dtu.dk}

Authors:
\itemize{
  \item Hjorleifur G Bergsteinsson \email{hgbe@dtu.dk}
}

}
\keyword{internal}
