% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineforecast-package.R
\docType{package}
\name{onlineforecast}
\alias{onlineforecast}
\title{Functions for online forecasting}
\description{
Functions for online forecasting
}
\details{
This package provides functions to for setting up forecast models which run in an online setting, e.g. like demand, solar and wind power forecasts updated regularly - often hourly, and forecasts up to 48 hours ahead.

See the website \url{https://onlineforecasting.org} for more information.
}
