% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{pbspline}
\alias{pbspline}
\title{Wrapper for \code{bspline} with \code{periodic=TRUE}}
\usage{
pbspline(
  X,
  Boundary.knots = NA,
  intercept = FALSE,
  df = NULL,
  knots = NULL,
  degree = 3,
  bknots = NA
)
}
\arguments{
\item{X}{see \code{?bspline}}

\item{Boundary.knots}{see \code{?bspline}}

\item{intercept}{see \code{?bspline}}

\item{df}{see \code{?bspline}}

\item{knots}{see \code{?bspline}}

\item{degree}{see \code{?bspline}}

\item{bknots}{see \code{?bspline}}
}
\description{
Wrapper for \code{bspline} with \code{periodic=TRUE}
}
\details{
Simply a wrapper.
}
\seealso{
Other Transform stage functions: 
\code{\link{bspline}()}
}
\concept{Transform stage functions}
