% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastmodel.R
\name{print.forecastmodel}
\alias{print.forecastmodel}
\title{Print forecast model}
\usage{
\method{print}{forecastmodel}(x, ...)
}
\arguments{
\item{x}{A forecastmodel object}

\item{...}{Not used.}
}
\description{
Prints a forecast model
}
\details{
A simple print out of the model output and inputs
}
