% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_to_message.R
\name{print_to_message}
\alias{print_to_message}
\title{Simple function for capturing from the print function and send it in a message().}
\usage{
print_to_message(...)
}
\arguments{
\item{...}{Passed to print which passed to message.}
}
\description{
Simple function for capturing from the print function and send it in a message().
}
