% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rls_update_cpp}
\alias{rls_update_cpp}
\title{Calculating k-step recursive least squares estimates}
\arguments{
\item{y}{Vector of observation}

\item{X}{Matrix of input variables (design matrix)}

\item{theta}{Vector of parameters (initial value)}

\item{P}{Covariance matrix (initial value)}

\item{lambda}{Forgetting factor}

\item{k}{Forecast horizon}

\item{n}{Length of the input}

\item{np}{Dimension of P (np x np)}

\item{istart}{Start index}

\item{kmax}{Keep only the last kmax rows for next time}
}
\description{
This function applies the k-step recursive least squares scheme to estimate
parameters in a linear regression model.
}
